+++ src/main/java/racingcar/constant/error/ErrorNameConstant.java
@@ -1,8 +1,9 @@
-package racingcar.constant;
+package racingcar.constant.error;
 
-public class ErrorConstant {
+public class ErrorNameConstant {
     public static final String ERROR_DUPLICATE_NAME = "중복된 자동차 이름이 존재합니다.";
     public static final String ERROR_LENGTH_NAME = "자동차 이름의 길이는 5 이하여야 합니다.";
     public static final String ERROR_EMPTY_NAME = "자동차 이름은 공백이 될 수 없습니다.";
     public static final String ERROR_DELIMITER_NAME = "입력값에 쉼표가 없습니다";
+    public static final String ERROR_MULTIPLE_NAME = "자동차는 한 대 이상이여야 합니다.";
 }
