--- src/main/java/racingcar/exception/EmptyNameException.java
@@ -1,9 +1,0 @@
-package racingcar.exception;
-
-import racingcar.constant.ErrorConstant;
-
-public class EmptyNameException extends IllegalArgumentException {
-    public EmptyNameException() {
-        super(ErrorConstant.ERROR_EMPTY_NAME);
-    }
-}
