--- src/main/java/racingcar/exception/LengthNameException.java
@@ -1,9 +1,0 @@
-package racingcar.exception;
-
-import racingcar.constant.ErrorConstant;
-
-public class LengthNameException extends IllegalArgumentException{
-    public LengthNameException() {
-        super(ErrorConstant.ERROR_LENGTH_NAME);
-    }
-}
