+++ src/main/java/racingcar/exception/input/name/EmptyNameException.java
@@ -1,0 +1,9 @@
+package racingcar.exception.input.name;
+
+import racingcar.constant.error.ErrorNameConstant;
+
+public class EmptyNameException extends IllegalArgumentException {
+    public EmptyNameException() {
+        super(ErrorNameConstant.ERROR_EMPTY_NAME);
+    }
+}
