+++ src/main/java/racingcar/exception/input/name/LengthNameException.java
@@ -1,0 +1,9 @@
+package racingcar.exception.input.name;
+
+import racingcar.constant.error.ErrorNameConstant;
+
+public class LengthNameException extends IllegalArgumentException{
+    public LengthNameException() {
+        super(ErrorNameConstant.ERROR_LENGTH_NAME);
+    }
+}
